Feature/the return of mon cli (#828)
* mon_cli (ftxui) renamed to mon_tui (text user interface)
* ecal_mon_cli (old cmdline version) added

+++ app/mon/mon_tui/src/tui/viewmodel/shell.hpp
@@ -141,7 +141,7 @@ public:
   {
     if(index < tabs.size())
     {
-      tab_selected = index;
+      tab_selected = static_cast<int>(index);
     }
   }
